Renamed to transaction(…), new transaction(…), protected to public
Renamed DatabaseModel.executeTransaction(…) to transaction(…).

Added variants of DatabaseModel.transaction(…) to match the full set of
transaction methods of the underlying Database.

Changed several methods of DatabaseModel from protected to public.  We
found that implementation were often having to override these methods to
make them package-accessible, so public is a better match.

1) getDatabase()
2) transaction(…)